src/main/java/com/maveric/digital/service/LocalDateConverter.java
package com.maveric.digital.service;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;
import java.util.Locale;
import java.util.Objects;

import com.opencsv.bean.AbstractBeanField;
import com.opencsv.exceptions.CsvConstraintViolationException;
import com.opencsv.exceptions.CsvDataTypeMismatchException;

public class LocalDateConverter extends AbstractBeanField {
	@Override
	protected Object convert(String value) throws CsvDataTypeMismatchException, CsvConstraintViolationException {
		if(Objects.nonNull(value) && !value.isBlank() && !value.equals("Project Start Date") && !value.equals("Project End Date")) {
			DateTimeFormatter formatter = DateTimeFormatter.ofPattern("dd-MM-yyyy", Locale.ENGLISH);
			return LocalDate.parse(value,formatter);
		}
		return null;
	}
	

}
